@@ -15,10 +15,10 @@
     "start": "ng serve",
     "build": "ng build",
     "build:prod": "npm run build -- --prod --aot",
-    "test": "ng test -sr",
-    "test:coverage": "rimraf coverage && npm run test -- -cc",
-    "lint": "ng lint --type-check",
-    "lint:fix": "ng lint --fix",
+    "test": "ng test",
+    "test:coverage": "rimraf coverage && npm run test -- --code-coverage",
+    "lint": "ng lint",
+    "lint:fix": "ng lint ngx-admin-demo --fix",
     "lint:styles": "stylelint ./src/**/*.scss",
     "lint:ci": "npm run lint && npm run lint:styles",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
@@ -30,15 +30,15 @@
   },
   "dependencies": {
     "@agm/core": "1.0.0-beta.2",
-    "@angular/animations": "~5.2.6",
-    "@angular/common": "~5.2.6",
-    "@angular/compiler": "~5.2.6",
-    "@angular/core": "~5.2.6",
-    "@angular/forms": "~5.2.6",
-    "@angular/http": "~5.2.6",
-    "@angular/platform-browser": "~5.2.6",
-    "@angular/platform-browser-dynamic": "~5.2.6",
-    "@angular/router": "~5.2.6",
+    "@angular/animations": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
     "@asymmetrik/ngx-leaflet": "3.0.1",
     "@nebular/auth": "^2.0.0-rc.7",
     "@nebular/security": "^2.0.0-rc.7",
@@ -65,20 +65,21 @@
     "normalize.css": "6.0.0",
     "pace-js": "1.0.2",
     "roboto-fontface": "0.8.0",
-    "rxjs": "5.5.5",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
     "socicon": "3.0.5",
     "tinymce": "4.5.7",
     "typeface-exo": "0.0.22",
     "web-animations-js": "2.2.5",
-    "zone.js": "0.8.18"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.1",
-    "@angular/compiler-cli": "~5.2.6",
-    "@angular/language-service": "~5.2.6",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/language-service": "6.0.0",
     "@compodoc/compodoc": "1.0.1",
-    "@types/googlemaps": "3.30.4",
     "@types/d3-color": "1.0.5",
+    "@types/googlemaps": "3.30.4",
     "@types/jasmine": "2.5.54",
     "@types/jasminewd2": "2.0.3",
     "@types/leaflet": "1.2.3",
@@ -101,6 +102,7 @@
     "ts-node": "3.2.2",
     "tslint": "5.7.0",
     "tslint-language-service": "0.9.6",
-    "typescript": "2.5.3"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.0"
   }
 }
