@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
 import { NavigationEnd, Router } from '@angular/router';
 import { Location } from '@angular/common';
-
-import { filter } from 'rxjs/operator/filter';
+import { filter } from 'rxjs/operators';
 
 declare const ga: any;
 
@@ -16,7 +15,9 @@ export class AnalyticsService {
 
   trackPageViews() {
     if (this.enabled) {
-      filter.call(this.router.events, (event) => event instanceof NavigationEnd)
+      this.router.events.pipe(
+        filter((event) => event instanceof NavigationEnd),
+      )
         .subscribe(() => {
           ga('send', {hitType: 'pageview', page: this.location.path()});
         });
